Added logic to handle fullpaths for bkg and font
If a full path is passed for either the background image or the font, this will bypass the "res" folder logic. Also, if the passed font fails to load, the program will perform a one-last attempt to load the default font on the res folder, if exists, in the executable path.